+++ docs/versionhistory.rst
@@ -33,6 +33,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - Bumped minimum versions of trio and curio to v0.16 and v1.2, respectively
 - Fixed a bug where a task group would abandon its subtasks if its own cancel scope was cancelled
   while it was waiting for subtasks to finish
+- Worked around curio's limitation where a task can only be cancelled twice (any cancellations
+  beyond that were ignored)
 
 **1.4.0**